Extract shared conference form parsing
create-conference and edit-conference duplicated ~40 identical lines
of milestone form parsing; parseConferenceForm is now the single
implementation. (Review section 4)
